requirements.txt
flask
python-kasa